requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
cachetools==5.3.3
catalogue==1.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dashscope==1.19.3
distro==1.9.0
einops==0.8.0
en_core_web_sm==2.3.0
exceptiongroup==1.2.1
filelock==3.15.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gevent==24.2.1
git-lfs==1.6
google-ai-generativelanguage==0.4.0
google-api-core==2.19.0
google-auth==2.30.0
google-generativeai==0.3.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
install==1.3.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.35.1
packaging==24.1
peft==0.11.1
pillow==10.3.0
pip==24.0
plac==1.1.3
preshed==3.0.9
progress==1.6
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge_score==0.1.2
rsa==4.9
safetensors==0.4.3
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
sympy==1.12.1
thinc==7.4.6
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.4
transformers==4.41.2
transformers-stream-generator==0.0.5
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==0.10.1
weasel==0.4.1
websocket==0.2.1
wheel==0.43.0
wrapt==1.16.0
xlwt==1.3.0
yarl==1.9.4
zhipuai==2.1.0.20240521
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
